requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
dj-rest-auth==6.0.0
Django==4.2.7
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-filter==24.2
django-leaflet==0.29.0
django-pwa==2.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
frozenlist==1.4.1
GDAL==3.4.1
idna==3.7
Kivy==2.3.0
Kivy-Garden==0.1.5
Markdown==3.6
multidict==6.0.5
numpy==1.26.2
pandas==2.2.3
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
twilio==9.2.3
typing_extensions==4.8.0
tzdata==2024.2
urllib3==2.2.2
yarl==1.9.4
